Refactor: 버튼 영역을 컴포넌트로 분리 및 디렉토리 이름 변경
